+++ lupa/tests/test_commands.py
@@ -134,7 +134,7 @@ class TestCheckConsistency(TestCase):
     @mock.patch.object(Command, 'printstatus')
     def test_process_data_error(self, _ptst, _ptnok):
         mdado = make(
-            'lupa.Dado',
+            'lupa.DadoEntidade',
             title='Bairro',
             id=3,
             database='DB',
@@ -189,7 +189,7 @@ class TestCheckConsistency(TestCase):
     def test_process_data(self, _excs, _ptst, _prok):
         command = Command()
         mdado = make(
-            'lupa.Dado',
+            'lupa.DadoEntidade',
             title='Bairro',
             id=3,
             database='DB',
@@ -227,7 +227,7 @@ class TestCheckConsistency(TestCase):
             name="Entidade Teste"
         )
         mdado = make(
-            'lupa.Dado',
+            'lupa.DadoEntidade',
             title='Bairro',
             id=3,
             database='DB',